+++ API/app/routers/text.py
@@ -1,6 +1,7 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 from ..db import get_text, get_texts, add_text, get_latest_text_of_user, delete_text, update_text
-from ..schemas import TextOut, TextCreate
+from ..schemas import TextOut, TextCreate, TokenData
+from ..oauth2 import get_current_user
 
 router = APIRouter(
     prefix="/texts",
@@ -37,7 +38,13 @@ def create_text(text: TextCreate):
 
 
 @router.delete("/{id}", response_model=TextOut, status_code=status.HTTP_200_OK)
-def remove_text(id: int):
+def remove_text(id: int, owner: TokenData = Depends(get_current_user)):
+    text = get_text(id)
+    if not text:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Text with id {id} not found")
+    if text.owner_id != owner.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"User not authorized to delete this text")
+    
     text = delete_text(id)
     if not text:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Text with id {id} not found")
@@ -45,7 +52,9 @@ def remove_text(id: int):
 
 
 @router.put("/{id}", response_model=TextOut, status_code=status.HTTP_200_OK)
-def change_text(id: int, text: TextCreate):
+def change_text(id: int, text: TextCreate, owner: TokenData = Depends(get_current_user)):
+    if text.owner_id != owner.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"User not authorized to change this text")
     text = update_text(id, text.title, text.content, text.owner_id, text.posted)
     if not text:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Text with id {id} not found")
